+++ src/black/__init__.py
@@ -816,8 +816,9 @@ def get_sources(
             if is_stdin:
                 path = Path(f"{STDIN_PLACEHOLDER}{path}")
 
-            if path.suffix == ".ipynb" and not jupyter_dependencies_are_installed(
-                warn=verbose or not quiet
+            if (
+                path.suffix == ".ipynb"
+                and not jupyter_dependencies_are_installed(warn=verbose or not quiet)
             ):
                 continue
 
@@ -936,8 +937,11 @@ def reformat_one(
             ):
                 if not cache.is_changed(src):
                     changed = Changed.CACHED
-            if changed is not Changed.CACHED and format_file_in_place(
-                src, fast=fast, write_back=write_back, mode=mode, lines=lines
+            if (
+                changed is not Changed.CACHED
+                and format_file_in_place(
+                    src, fast=fast, write_back=write_back, mode=mode, lines=lines
+                )
             ):
                 changed = Changed.YES
             if cache is not None and (
@@ -1422,8 +1426,9 @@ def get_features_used(
             features.add(Feature.LAZY_IMPORTS)
 
         elif n.type == syms.decorator:
-            if len(n.children) > 1 and not is_simple_decorator_expression(
-                n.children[1]
+            if (
+                len(n.children) > 1
+                and not is_simple_decorator_expression(n.children[1])
             ):
                 features.add(Feature.RELAXED_DECORATORS)
 
@@ -1481,8 +1486,9 @@ def get_features_used(
         elif n.type == syms.match_stmt:
             features.add(Feature.PATTERN_MATCHING)
 
-        elif n.type in {syms.subscriptlist, syms.trailer} and any(
-            child.type == syms.star_expr for child in n.children
+        elif (
+            n.type in {syms.subscriptlist, syms.trailer}
+            and any(child.type == syms.star_expr for child in n.children)
         ):
             features.add(Feature.VARIADIC_GENERICS)