+++ src/black/nodes.py
@@ -931,15 +931,18 @@ def is_import(leaf: Leaf) -> bool:
 
 def is_with_or_async_with_stmt(leaf: Leaf) -> bool:
     """Return True if the given leaf starts a with or async with statement."""
-    return bool(
-        leaf.type == token.NAME
-        and leaf.value == "with"
-        and leaf.parent
-        and leaf.parent.type == syms.with_stmt
-    ) or bool(
-        leaf.type == token.ASYNC
-        and leaf.next_sibling
-        and leaf.next_sibling.type == syms.with_stmt
+    return (
+        bool(
+            leaf.type == token.NAME
+            and leaf.value == "with"
+            and leaf.parent
+            and leaf.parent.type == syms.with_stmt
+        )
+        or bool(
+            leaf.type == token.ASYNC
+            and leaf.next_sibling
+            and leaf.next_sibling.type == syms.with_stmt
+        )
     )
 
 
@@ -974,17 +977,19 @@ def is_type_ignore_comment(leaf: Leaf, mode: Mode) -> bool:
     """Return True if the given leaf is a type comment with ignore annotation."""
     t = leaf.type
     v = leaf.value
-    return t in {token.COMMENT, STANDALONE_COMMENT} and is_type_ignore_comment_string(
-        v, mode
+    return (
+        t in {token.COMMENT, STANDALONE_COMMENT}
+        and is_type_ignore_comment_string(v, mode)
     )
 
 
 def is_type_ignore_comment_string(value: str, mode: Mode) -> bool:
     """Return True if the given string match with type comment with
     ignore annotation."""
-    return is_type_comment_string(value, mode) and value.split(":", 1)[
-        1
-    ].lstrip().startswith("ignore")
+    return (
+        is_type_comment_string(value, mode)
+        and value.split(":", 1)[1].lstrip().startswith("ignore")
+    )
 
 
 def wrap_in_parentheses(parent: Node, child: LN, *, visible: bool = True) -> None: